requirements.txt
numpy
pandas
tensorflow-macos
scikit-learn
matplotlib
yfinance
streamlit
